fix(Buttons): Merged the two button components
