+++ src/components/button/button.indigo.scss
@@ -5,8 +5,7 @@
     $m-size: utils.rem(32px);
     $s-size: utils.rem(24px);
 
-    igc-button[size='large']::part(base),
-    igc-link-button[size='large']::part(base) {
+    igc-button[size='large']::part(base) {
         --size: #{$l-size};
         --igc-border-radius: calc(var(--size) / 2);
         border-radius: var(--igc-border-radius);
@@ -17,8 +16,7 @@
         }
     }
 
-    igc-button[size='medium']::part(base),
-    igc-link-button[size='medium']::part(base) {
+    igc-button[size='medium']::part(base) {
         --size: #{$m-size};
         --igc-border-radius: calc(var(--size) / 2);
         border-radius: var(--igc-border-radius);
@@ -29,8 +27,7 @@
         }
     }
 
-    igc-button[size='small']::part(base),
-    igc-link-button[size='small']::part(base) {
+    igc-button[size='small']::part(base) {
         --size: #{$s-size};
         --igc-border-radius: calc(var(--size) / 2);
         border-radius: var(--igc-border-radius);
@@ -41,8 +38,7 @@
         }
     }
 
-    igc-button[variant='flat']:not([disabled])::part(base),
-    igc-link-button[variant='flat']:not([disabled])::part(base) {
+    igc-button[variant='flat']:not([disabled])::part(base) {
         color: #{utils.color(gray, 700)};
 
         transition: color .15s ease-in-out, box-shadow .15s ease-in-out;
@@ -68,8 +64,7 @@
         }
     }
 
-    igc-button[variant='outlined']:not([disabled])::part(base),
-    igc-link-button[variant='outlined']:not([disabled])::part(base) {
+    igc-button[variant='outlined']:not([disabled])::part(base) {
         color: #{utils.color(gray, 700)};
         transition: color .15s ease-out;
         box-shadow: inset 0 0 0 2px #{utils.color(gray, 500)};
@@ -104,10 +99,8 @@
         }
     }
 
-    igc-button[variant='fab']:not([disabled])::part(base),
-    igc-link-button[variant='fab']:not([disabled])::part(base),
-    igc-button[variant='contained']:not([disabled])::part(base),
-    igc-link-button[variant='contained']:not([disabled])::part(base) {
+    igc-button[variant='fab']:not([disabled])::part(base)
+    igc-button[variant='contained']:not([disabled])::part(base) {
         color: #{utils.contrast-color(primary, 600)};
 
         &:hover {
@@ -134,33 +127,25 @@
     }
 
     igc-button[variant='fab']:not([disabled])::part(base),
-    igc-link-button[variant='fab']:not([disabled])::part(base),
-    igc-button[variant='fab'][disabled]::part(base),
-    igc-link-button[variant='contained'][disabled]::part(base) {
+    igc-button[variant='fab'][disabled]::part(base) {
         border-radius: calc(var(--igc-border-radius) * 100);
     }
 
     igc-button[variant='contained']::part(base),
-    igc-button[variant='fab']::part(base),
-    igc-link-button[variant='contained']::part(base),
-    igc-link-button[variant='fab']::part(base) {
+    igc-button[variant='fab']::part(base) {
         box-shadow: var(--igc-elevation-0);
     }
 
     // DISABLED START
     igc-button[variant='flat'][disabled]::part(base),
-    igc-link-button[variant='flat'][disabled]::part(base),
-    igc-button[variant='outlined'][disabled]::part(base),
-    igc-link-button[variant='outlined'][disabled]::part(base) {
+    igc-button[variant='outlined'][disabled]::part(base) {
         background: transparent;
         box-shadow: none;
         color: #{utils.color(gray, 400)};
     }
 
     igc-button[variant='fab'][disabled]::part(base),
-    igc-button[variant='contained'][disabled]::part(base),
-    igc-link-button[variant='fab'][disabled]::part(base),
-    igc-link-button[variant='contained'][disabled]::part(base) {
+    igc-button[variant='contained'][disabled]::part(base) {
         background: #{utils.color(gray, 100)};
         color: #{utils.color(gray, 400)};
     }
