+++ stories/button.stories.ts
@@ -16,6 +16,32 @@ const metadata = {
         type: 'inline-radio',
       },
     },
+    href: {
+      type: 'string',
+      description: 'The URL the button points to.',
+      control: 'text',
+    },
+    download: {
+      type: 'string',
+      description:
+        'Prompts to save the linked URL instead of navigating to it.',
+      control: 'text',
+    },
+    target: {
+      type: '"_blank" | "_parent" | "_self" | "_top" | undefined',
+      description:
+        'Where to display the linked URL, as the name for a browsing context.',
+      options: ['_blank', '_parent', '_self', '_top', 'undefined'],
+      control: {
+        type: 'select',
+      },
+    },
+    rel: {
+      type: 'string',
+      description:
+        'The relationship of the linked URL.\nSee https://developer.mozilla.org/en-US/docs/Web/HTML/Link_types',
+      control: 'text',
+    },
     disabled: {
       type: 'boolean',
       description: 'Determines whether the button is disabled.',
@@ -45,13 +71,17 @@ const metadata = {
 export default metadata;
 interface ArgTypes {
   type: 'button' | 'reset' | 'submit';
+  href: string;
+  download: string;
+  target: '_blank' | '_parent' | '_self' | '_top' | undefined;
+  rel: string;
   disabled: boolean;
   variant: 'flat' | 'contained' | 'outlined' | 'fab';
   size: 'small' | 'medium' | 'large';
 }
 // endregion
 
-const Template: Story<ArgTypes, Context> = (
+const ButtonTemplate: Story<ArgTypes, Context> = (
   { disabled = false, size, variant, type }: ArgTypes,
   { globals: { direction } }: Context
 ) => {
@@ -70,4 +100,35 @@ const Template: Story<ArgTypes, Context> = (
   `;
 };
 
-export const Basic = Template.bind({});
+const LinkTemplate: Story<ArgTypes, Context> = (
+  {
+    disabled = false,
+    size,
+    variant,
+    href = 'http://www.infragistics.com',
+    download,
+    rel,
+    target,
+  }: ArgTypes,
+  { globals: { direction } }: Context
+) => html`
+  <igc-button
+    .disabled=${disabled}
+    .size=${size}
+    .variant=${variant}
+    .href=${href}
+    .download=${download}
+    .rel=${rel}
+    .target=${target}
+    dir=${ifDefined(direction)}
+  >
+    Click me
+  </igc-button>
+`;
+
+export const Button = ButtonTemplate.bind({});
+export const Link = LinkTemplate.bind({});
+export const BlankTarget = LinkTemplate.bind({});
+BlankTarget.args = {
+  target: '_blank',
+};
